IBM Data Science - Course(7): Data Analysis Final
Instructions
In this assignment, you are a Data Analyst working at a Real Estate Investment Trust. The Trust would like to start investing in Residential real estate. You are tasked with determining the market price of a house given a set of features. You will analyze and predict housing prices using attributes or features such as square footage, number of bedrooms, number of floors, and so on. This is a template notebook; your job is to complete the ten questions
I will complete the required project in this notebook in a detailed way to make it educational and usable as a review notebook later on.